fuse: Use DirectMountStrict over DirectMount
Currently, when `direct_mount` is set in the configuration, `fusermount`
may still be used if direct mounting fails. While this may be useful
from a configuration ease perspective, it squelches the error from
direct mounting, making the issue harder to fix, and disobeys the
configuration intent (which presumably was to avoid using `fusermount`).

This change fixes the issue by setting `DirectMountStrict` instead of
`DirectMount` to all calls to `go_fuse` when creating FUSE mounts.

Tested: not yet